Moving MoveToFort to the top level (#1605)
* Moving MoveToFort to the top level

* Fixing bad import
